requirements.txt
streamlit
pandas
scikit-surprise
numpy
scikit-learn
tensorflow
keras
matplotlib
seaborn
